Ensure chord sheet white space is parsed correctly
Example:

C                G
Whisper words of wisdom

There are 15 spaces between the C and G chord. Parsing this will return
the first pair as a C and 14 spaces. Keeping the white space is
important when the line ends with some chords having multiple spaces in
between, like this:

C                G              F  C/E Dm          C
Whisper words of wisdom, let it be

The white space between Dm and C is important for formatting purposes.