Implement ChordPro annotations (#1127)
Quoted from the ChordPro docs
(https://www.chordpro.org/chordpro/chordpro-introduction/#markup-and-annotations):

Annotations are textual remarks placed above the lyrics, just like chords.
Annotations are specified with [*text], again just like chords. Depending
on the software used to process the ChordPro data, annotations may be
rendered in an outstanding manner.

This is a first iteration on annotation support, where the annotation is
added to a `ChordLyricsPair`. The pair can either have `chords` or
`annotation`.

In `TextFormatter` and `ChordsOverWordsFormatter` annotations are
treated equally to lyrics. In `HtmlDivFormatter` and
`HtmlTableFormatter` annotatioins are rendered similar to how lyrics are
rendered, but the wrapping element is given the css class `annotation`
instead of `lyrics`.

See https://www.chordpro.org/chordpro/chordpro-introduction/#markup-and-annotations

Resolves #980