Full support for sections
Full support for tab, abc, ly (Lilypond) and grid sections.

For example, when encountering a section like:

    {start_of_tab}
    Tab line 1
    Tab line 2
    {end_of_tab}

the contents of the tab section will be parsed as two separate lines
having one item: a `Literal` with it's `string` being the line contents
(without trailing newline).

That also means, `Line`.`items` can not only be `ChordLyricsPair`,
`Comment`, `Tag` or `Ternary`, but also `Literal`.

The is a new template helper called `isLiteral` to check for literals
when rendering a template.

See also:
- https://www.chordpro.org/chordpro/directives-env_tab/
- https://chordpro.org/chordpro/directives-env_abc/
- https://chordpro.org/chordpro/directives-env_ly/
- https://chordpro.org/chordpro/directives-env_grid/