Return null for missing metadata
The return value for missing metadata was not consistent. According to
the type docs `Metadata#getSingle()` would always return a `string`.
However, when the metadata entry was missing it would actually return
`undefined`.

Explicitly returning `undefined` is not recommended. This PR fixes the
type annotation and changes the value for 'none' to `null`.

Related to #1161
Thanks to @ChrisMBarr for reporting