Make decapo configurable
When the song has a capo, decapo will transpose the formatted song to
eliminate the capo. This used to be the default and was not
configurable. This change makes it configurable and disables it by
default.

You can explicitly enable it by passing `{ decapo: true }` to the
constructor formatter, eg:

```ts
new HtmlDivFormatter({ decapo: true }).format(song)
```

Thanks to @edonv for submitting the issue.

Resolves https://github.com/martijnversluis/ChordSheetJS/issues/1447

See also script/generate_chord_definitions.ts